+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Examples/Servo/servoExampleButtonControl.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Examples;
+package org.firstinspires.ftc.teamcode.Examples.Servo;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
